new function: allow category be edited.

+++ app/src/main/java/com/just/print/ui/fragment/ConfigMenuFragment.java
@@ -24,7 +24,6 @@ import com.just.print.ui.holder.AddPrintViewHolder;
 import com.just.print.ui.holder.ConfigCategoryViewHolder;
 import com.just.print.ui.holder.ConfigMenuViewHolder;
 import com.stupid.method.adapter.IXOnItemClickListener;
-import com.stupid.method.adapter.IXOnItemLongClickListener;
 import com.stupid.method.adapter.XAdapter2;
 import com.stupid.method.reflect.StupidReflect;
 import com.stupid.method.reflect.annotation.XClick;
@@ -39,9 +38,10 @@ import static android.content.DialogInterface.OnClickListener;
 /**
  * Created by wangx on 2016/11/2.
  */
-public class ConfigMenuFragment extends BaseFragment implements IXOnItemLongClickListener {
+public class ConfigMenuFragment extends BaseFragment{//} implements IXOnItemLongClickListener {
 
     private static Menu modifyingMenu = null;
+    private Category modifyingCategory = null;
 
     XAdapter2<Category> categoryXAdapter;
     @XViewByID(R.id.lvCategory)
@@ -51,6 +51,8 @@ public class ConfigMenuFragment extends BaseFragment implements IXOnItemLongClic
     private TextView tvMenuTitle;
     @XViewByID(R.id.lvMenu)
     private ListView lvMenu;
+
+    MenuDao menuDao;
     private XAdapter2<Menu> menuXAdapter;
     IXOnItemClickListener categoryAdapterClick = new IXOnItemClickListener() {
         @Override
@@ -59,10 +61,16 @@ public class ConfigMenuFragment extends BaseFragment implements IXOnItemLongClic
                 case android.R.id.content:
                     onChangeCategory(categoryXAdapter.get(i));
                     break;
+
+                case R.id.modifyCategory:
+                    //show up hte modification dialog.
+                    modifyingCategory = categoryXAdapter.get(i);
+                    showCategory();
+                    TextView tvCategory = (TextView) findViewById(R.id.tvCategory);
+                    tvCategory.setText(modifyingCategory.getCname());
+
+                    break;
                 case R.id.delCategory:
-                    /**
-                     * 删除菜品种类
-                     */
                     Category cag = categoryXAdapter.get(i);
                     long cid = cag.getId();
                     List<Menu> mList = menuDao._queryCategory_MenuList(cid);
@@ -90,6 +98,8 @@ public class ConfigMenuFragment extends BaseFragment implements IXOnItemLongClic
                     modifyingMenu = menuXAdapter.get(i);
                     ((EditText)findViewById(R.id.etMenuID)).setText(menuXAdapter.get(i).getID());
                     ((EditText)findViewById(R.id.etMenu)).setText(menuXAdapter.get(i).getMname());
+                    Double price = menuXAdapter.get(i).getPrice();
+                    ((EditText)findViewById(R.id.price)).setText(price == null ? "" : price.toString());
 
                     addPrints = new ArrayList<Printer>();
                     for(M2M_MenuPrint m2m_p : menuXAdapter.get(i).getM2M_MenuPrintList()){
@@ -134,6 +144,7 @@ public class ConfigMenuFragment extends BaseFragment implements IXOnItemLongClic
             }
         }
     };
+
     private void onChangeCategory(Category category) {
         mCategory = category;
         if (category != null) {
@@ -147,16 +158,13 @@ public class ConfigMenuFragment extends BaseFragment implements IXOnItemLongClic
         return R.layout.config_menu_fragment;
     }
 
-    MenuDao menuDao;
-
-
     @Override
     public void onCreated(Bundle savedInstanceState) {
         new StupidReflect(this, getView()).init();
 
         categoryXAdapter = new XAdapter2<Category>(getContext(), ConfigCategoryViewHolder.class);
         categoryXAdapter.setClickItemListener(categoryAdapterClick);
-        categoryXAdapter.setLongClickItemListener(this);
+        //categoryXAdapter.setLongClickItemListener(this);
         lvCategory.setAdapter(categoryXAdapter);
 
         menuXAdapter = new XAdapter2<Menu>(getContext(), ConfigMenuViewHolder.class);
@@ -204,23 +212,34 @@ public class ConfigMenuFragment extends BaseFragment implements IXOnItemLongClic
     @XClick({R.id.addMenu})
     private void addMenu(@XGetValueByView(fromId = R.id.etMenuID) TextView mid,
                          @XGetValueByView(fromId = R.id.etMenu) TextView mname,
+                         @XGetValueByView(fromId = R.id.price) TextView mprice,
                          @XGetValueByView(fromId = R.id.selectPrint) TextView print) {
 
         if (mCategory == null) {
             showToast("Please choose a category");
             return;
         }
-        Menu menu = menuDao.load(mid.getText().toString());
-        if (mid.getText().length() == 0 || mname.getText().length() == 0) {
+
+        if (mid.getText().length() == 0 || mname.getText().length() == 0 || mprice.getText().length() == 0 ) {
             showToast("Please input " + (mid.getText().length() == 0 ? "code" : "name"));
             return;
         }
+        Double price = 0.0;
+        try{
+            price =Double.valueOf(String.valueOf(mprice.getText()));
+        } catch(Exception e){
+            showToast("Please input valid price. e.g. 15.99");
+            return;
+        }
+
         if (addPrints == null || addPrints.size() == 0) {
             showToast("Please select printer");
             return;
         }
 
+
         //filter out "is adding" and "code is ocupied" case.
+        Menu menu = menuDao.load(mid.getText().toString());
         if (menu != null && menu.getState() == State.def && modifyingMenu == null) {
             showToast("the code is used already");
             return;
@@ -235,21 +254,23 @@ public class ConfigMenuFragment extends BaseFragment implements IXOnItemLongClic
         }
         */
 
-        if(menu == null){   //is adding mode( worry free about code duplication with existing one, it's filtered out.)
+        if(menu == null){   //the code mapping to no record in db.
             menu = new Menu();
             menu.setID(mid.getText().toString().trim());
             menu.setState(State.def);
             menu.setMname(mname.getText().toString().trim());
             menu.setCid(mCategory.getId());
+            menu.setPrice(price);
             menuDao.insert(menu);
-        }else{              //is modifying mode. means duplicated with others. but it's modifying, (duplicated and not modifying case has been filtered out).
+        }else{              //the code is same with some record in db.)
             menu.setID(mid.getText().toString().trim());
             menu.setState(State.def);
             menu.setMname(mname.getText().toString().trim());
             menu.setCid(mCategory.getId());
+            menu.setPrice(price);
             menuDao.update(menu);
         }
-        if(modifyingMenu != null && !modifyingMenu.getID().equals(menu.getID())){
+        if(modifyingMenu != null && !modifyingMenu.getID().equals(menu.getID())){ //if it's modifying, and the code is changed to be same with some other existing one.
             menuDao.delete(modifyingMenu);
         }
 
@@ -270,7 +291,6 @@ public class ConfigMenuFragment extends BaseFragment implements IXOnItemLongClic
         modifyingMenu = null;
         findViewById(R.id.menuLayout).setVisibility(View.GONE);
         loadMenu();
-
     }
 
     /**
@@ -284,16 +304,16 @@ public class ConfigMenuFragment extends BaseFragment implements IXOnItemLongClic
         }
 
         Category cat = null;
-        if (mModifyCategoryTemp == null)
+        if (modifyingCategory == null)
             cat = new Category();
-        else cat = mModifyCategoryTemp;
+        else cat = modifyingCategory;
         cat.setCname(cname.getText().toString().trim());
         cat.setState(State.def);
         Applic.app.getDaoMaster().newSession().getCategoryDao().insertOrReplace(cat);
         cat.updateAndUpgrade();
         loadCategory();
         cname.setText("");
-        mModifyCategoryTemp = null;
+        modifyingCategory = null;
         findViewById(R.id.categoryLayout).setVisibility(View.GONE);
     }
 
@@ -359,15 +379,13 @@ public class ConfigMenuFragment extends BaseFragment implements IXOnItemLongClic
         return "Menu Management";
     }
 
-    Category mModifyCategoryTemp;
-
-    @Override
-    public boolean onLongClickItem(View view, int i) {
-        mModifyCategoryTemp = categoryXAdapter.get(i);
-        showCategory();
-        TextView tvCategory = (TextView) findViewById(R.id.tvCategory);
-        tvCategory.setText(mModifyCategoryTemp.getCname());
-
-        return true;
-    }
+//    @Override
+//    public boolean onLongClickItem(View view, int i) {
+//        modifyingCategory = categoryXAdapter.get(i);
+//        showCategory();
+//        TextView tvCategory = (TextView) findViewById(R.id.tvCategory);
+//        tvCategory.setText(modifyingCategory.getCname());
+//
+//        return true;
+//    }
 }
